[rilbinder] Fix logging. JB#44551
ofono's DBG magic doesn't work in a shared library because
debug descriptors end up in library's own __debug section
and they never get enabled because ofono doesn't know anything
about them.

Expose libglibutil style descriptors instead and explicitly
hook them up.
